Split widget: signal row
